9 dec/1

+++ app/Livewire/Forms/EventForm.php
@@ -2,19 +2,19 @@
 
 namespace App\Livewire\Forms;
 
+use App\Models\Contact;
 use App\Models\Event;
-use App\Models\Participant;
-use App\Models\Presentation;
 use App\Models\Project;
+use App\Models\User;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Gate;
-use Illuminate\Support\Facades\Redirect;
-use Livewire\Attributes\Computed;
 use Livewire\Attributes\Validate;
 use Livewire\Form;
 
 class EventForm extends Form
 {
+    use AuthorizesRequests;
+
     public ?Event $event;
 
     #[Validate('required|min:3')]
@@ -26,7 +26,6 @@ class EventForm extends Form
     public $duration;
 
     public $status;
-
 
     public function setEvent(Event $event)
     {
@@ -43,10 +42,18 @@ class EventForm extends Form
 
     public function addProject($project_id, $weight)
     {
+        $this->authorize('handleProject', [$this->event, Project::find($project_id)]);
+
+        if ($weight < 0 || $weight === null) {
+            $weight = 1;
+        }
+
         $this->event->projects()->attach($project_id, ['weight' => $weight]);
 
-        foreach ($this->event->students as $student) {
-            $student->projects()->attach($project_id);
+        if (isset($this->event->students)) {
+            foreach ($this->event->students as $student) {
+                $student->projects()->attach($project_id, ['event_id' => $this->event->id]);
+            }
         }
 
         session()->now('success', __('events.project_added'));
@@ -54,26 +61,36 @@ class EventForm extends Form
 
     public function removeProject($project_id)
     {
+        $this->authorize('update', $this->event);
+
         $this->event->projects()->detach($project_id);
     }
 
-   public function addEvaluator($evaluator_id)
+    public function addEvaluator($evaluator_id)
     {
+        $this->authorize('handleContact', [$this->event, Contact::find($evaluator_id)]);
+
         $this->event->evaluators()->attach($evaluator_id, ['role' => 'evaluator']);
     }
 
     public function removeEvaluator($evaluator_id)
     {
+        $this->authorize('update', $this->event);
+
         $this->event->evaluators()->detach($evaluator_id);
     }
 
     public function addStudent($student_id)
     {
+        $this->authorize('handleContact', [$this->event, Contact::find($student_id)]);
+
         $this->event->students()->attach($student_id, ['role' => 'student']);
     }
 
     public function removeStudent($student_id)
     {
+        $this->authorize('update', $this->event);
+
         $this->event->students()->detach($student_id);
     }
 
@@ -90,11 +107,16 @@ class EventForm extends Form
         $this->event = Auth::user()
             ->events()
             ->save(new Event($this->only('name', 'date')));
-
     }
 
     public function update()
     {
+        $this->authorize('update', $this->event);
+
+        if ($this->event->status !== null) {
+            session()->flash('error', __('error.event_over'));
+            return;
+        }
         $this->validate();
 
         $this->event->update($this->all());
@@ -102,6 +124,8 @@ class EventForm extends Form
 
     public function destroy()
     {
+        $this->authorize('delete', $this->event);
+
         $this->event->delete();
     }
 }
